Adjusted the Forces to move and jump, also refactored the camera controller to be the new camera look target controller
The camera look target controller is a vector position which rotates around the player 1 unit infront over a sphere. The cinemachine virtual camera looks at this so we can pan the view up and down as well as left and right. The rotation now needs to be handled by the player controller physicis which is called by various states.