added better camera look trage script which does not tear the transform object (player)
Adjusted the settings to the follow cam, and updated the look target script which rotates the object as an orbital around the parent of the look target vector. **NOTE -- we need to add a clamp to this, as it will continously follow the spherical path around the player. Also added in an exception check to make sure we have the player set.